Build JS with Grunt not Brunch
I decided to switch to Grunt, because it's easy to prefix a license text to a
JS or CSS bundle, with Grunt, but it seems annoyingly complicated with Brunch
(I would have to find out how to tell the minifier not to remove it).

And, more importantly, with Grunt I can have both .js and .min.js configs in
the same file. That wasn't possible with Brunch (although the Brunch devs are
supposedly planning to fix this).